% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop-tif.R
\name{crop_global_tif}
\alias{crop_global_tif}
\title{crop_global_tif}
\usage{
crop_global_tif(city = "bristol", expand = 0.1)
}
\arguments{
\item{city}{Name of city (one of "accra", "kathmandu", or "bristol")}

\item{expand}{Relative factor by which to expand cropping limits beyond
rectangular bounding box of city}
}
\value{
Nothing Files are written to the corresponding directories
}
\description{
There are no worldpop data for Europe, but the EC Joint Research Centre Data
Catalogue offers 250m resolution global population density \code{tif} files.
These are huge, and presumed to be downloaded into the root directory of this
project. This function crops them to designated cities (actuall only Bristol
at present) and deposits the results in the appropriate \code{[city]/popdens}
directory.
}
\details{
The main JRCDC page is
\url{http://data.jrc.ec.europa.eu/dataset/jrc-ghsl-ghs_pop_gpw4_globe_r2015a},
leading to the download of GHS_POP_GPW42015_GLOBE_R2015A_54009_250_v1_0.zip,
which is 1GB, and is not archived in this repo.
}
\note{
This should be run from somewhere within the \code{whodata} repository
structure; anywhere else will likely fail. It also takes a r-e-a-l-l-y long
time because of the raster re-projection.
}
