% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop.R
\name{download_bristol_pop}
\alias{download_bristol_pop}
\title{download_bristol_pop
Download the \code{tif} file disected with \link{crop_global_tif}}
\usage{
download_bristol_pop()
}
\description{
download_bristol_pop
Download the \code{tif} file disected with \link{crop_global_tif}
}
