% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_popdens_nodes}
\alias{download_popdens_nodes}
\title{download_popdens_nodes}
\usage{
download_popdens_nodes(city = NULL)
}
\arguments{
\item{city}{Default uploads both cities, otherwise one of Accra or Kathmandu.}
}
\description{
download_popdens_nodes
}
