% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_who_data}
\alias{download_who_data}
\title{download_who_data}
\usage{
download_who_data()
}
\description{
Download all WHO data from github repo via \pkg{piggyback}. This function can
be used both for initial download, and for periodic refreshment of data. Only
those data which have been updated will be downloaded anew.
}
