% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places-goog.R
\name{get_accra_gplaces}
\alias{get_accra_gplaces}
\title{get_accra_gplaces}
\usage{
get_accra_gplaces(radius = 200)
}
\arguments{
\item{radius}{Radius to be used for googel nearby place search (in metres)}
}
\value{
A \code{data.frame} of place locations, names, and types
}
\description{
Get list of places according to googel
}
\note{
This function requires setting a personal key for the google API with
\code{Sys.setenv(GOOGLE_API = "<my_key>")}.
}
