% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm-piggyback.R
\name{osm_files}
\alias{osm_files}
\title{osm_files}
\usage{
osm_files()
}
\value{
Character vector of file names to be used to extract files from
github releases via \pkg{piggyback}
}
\description{
Get lists of OSM files
}
