% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop.R
\name{upload_bristol_pop}
\alias{upload_bristol_pop}
\title{upload_bristol_pop
Upload the \code{tif} file disected with \link{crop_global_tif}}
\usage{
upload_bristol_pop()
}
\description{
upload_bristol_pop
Upload the \code{tif} file disected with \link{crop_global_tif}
}
