% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{upload_flows}
\alias{upload_flows}
\title{upload_flows}
\usage{
upload_flows(city = NULL, overwrite = TRUE)
}
\arguments{
\item{city}{Default uploads both cities, otherwise one of Accra or Kathmandu.}

\item{overwrite}{Should generally be set to \code{TRUE}.}
}
\description{
upload_flows
}
