% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_bp}
\alias{who3_bp}
\title{who3_bp}
\usage{
who3_bp(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\value{
Matrix of coordinates of bounding polygon
}
\description{
get bounding polygons for WHO3 cities (Accra, Kathmandu)
}
