% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_buildings}
\alias{who3_buildings}
\title{who3_buildings}
\usage{
who3_buildings(city, save = TRUE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\description{
who3_buildings
}
