% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_bus_centrality}
\alias{who3_bus_centrality}
\title{who3_bus_centrality}
\usage{
who3_bus_centrality(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\description{
Calculate centrality of bus networks for WHO3 cities, and assign vertex
centrality values to each bus stop, in order to use these as proxies for
numbers of passengers.
}
