% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_bus_network}
\alias{who3_bus_network}
\title{who3_bus_network}
\usage{
who3_bus_network(city)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}
}
\description{
Get bus network data for WHO3 cities (Accra, Kathmandu).
}
