% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_bus_stops}
\alias{who3_bus_stops}
\title{who3_bus_stops}
\usage{
who3_bus_stops(city, save = TRUE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}
}
\description{
Get bus stop data for WHO3 cities (Accra, Kathmandu).
}
