% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_centrality}
\alias{who3_centrality}
\title{who3_centrality}
\usage{
who3_centrality(city, save = TRUE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
A \pkg{dodgr} representation of the street network with a centrality
column calculated from time-based weighting for motorcar transport.
}
\description{
who3_centrality
}
