% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_disperse_centrality}
\alias{who3_disperse_centrality}
\title{who3_disperse_centrality}
\usage{
who3_disperse_centrality(city, disperse_width = 200)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{disperse_width}{Width in metres defining Gaussian dispersal of
vehicular emissions.}
}
\description{
Take a network with centrality column, disperse values away from network
edges, and map onto pedestrian network.
}
