% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who3.R
\name{who3_network}
\alias{who3_network}
\title{Get street network for WHO# cities (Accra and Kathmandu)}
\usage{
who3_network(city, save = FALSE, quiet = FALSE)
}
\arguments{
\item{city}{Name of city for which to obtain bounding polygon}

\item{save}{If `TRUE`, save data to internal `who-data` directory.}

\item{quiet}{If `FALSE`, display progress information on screen.}
}
\value{
SC-formatted street network for nominated city.
}
\description{
Get street network for WHO# cities (Accra and Kathmandu)
}
