% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldpop.R
\name{worldpop_files}
\alias{worldpop_files}
\title{worldpop_files}
\usage{
worldpop_files(type = "zip")
}
\arguments{
\item{type}{either "zip" (pre-processed) or "tif" (processed)}
}
\value{
Character vector of file names to be used to extract files from
github releases via \pkg{piggyback}
}
\description{
Get lists of worldpop files
}
